+++ src/dnt.ts
@@ -89,7 +89,7 @@ await build({
       {
         package: {
           name: "@js-temporal/polyfill",
-          version: "^0.4.4",
+          version: "^0.5.0",
         },
         globalNames: [
           {