+++ CHANGES.md
@@ -73,6 +73,9 @@ To be released.
  -  Updated *uri-template-router* to 0.0.17 which fixes bundler errors on
     Rollup.  [[#221]]
 
+ -  Improved error handling and logging for document loader when KV store
+    operations fail.  [[#223] by Revath S Kumar]
+
  -  Fixed a bug of the `fedify inbox` command where it had failed to render
     the web interface when the `fedify` command was installed using
     `deno install` command from JSR.
@@ -98,6 +101,7 @@ To be released.
 [#215]: https://github.com/fedify-dev/fedify/pull/215
 [#220]: https://github.com/fedify-dev/fedify/issues/220
 [#221]: https://github.com/fedify-dev/fedify/issues/221
+[#223]: https://github.com/fedify-dev/fedify/pull/223
 [multibase]: https://github.com/multiformats/js-multibase